chore: migrate worker scripts to typescript
